Manual views now work.
Links in the manual list went directly to a website that's seemingly discontinued instead of going to the manual_view page like it should. Now that works but ran into a problem where on the site it looked for the variable $type, which was seemingly meant to call the manuals table in the database. However, it never sent the $type variable to this page, since it has since been changed to $manual.